backend/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-ai-projects
azure-ai-agents
azure-ai-documentintelligence
openpyxl
python-dotenv
playwright
